+++ packages/block-editor/src/components/line-height-control/index.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import classNames from 'classnames';
-
-/**
  * WordPress dependencies
  */
 import deprecated from '@wordpress/deprecated';
@@ -93,29 +88,34 @@ export function LegacyLineHeightControl( { value: lineHeight, onChange } ) {
 export default function LineHeightControl( {
 	value: lineHeight,
 	onChange,
-	__unstableHasLegacyStyles = true,
+	__nextHasNoMarginBottom = false,
+	__unstableInputWidth = '60px',
 } ) {
 	const isDefined = isLineHeightDefined( lineHeight );
 	const value = isDefined ? lineHeight : RESET_VALUE;
 
-	if ( __unstableHasLegacyStyles ) {
-		deprecated( 'Legacy styles for wp.blockEditor.LineHeightControl', {
-			since: '5.10',
-			hint:
-				'Set the `__unstableHasLegacyStyles` prop to false to start opting into the new styles, which will become the default in a future version.',
-		} );
+	if ( ! __nextHasNoMarginBottom ) {
+		deprecated(
+			'Bottom margin styles for wp.blockEditor.LineHeightControl',
+			{
+				since: '6.0',
+				version: '6.4',
+				hint:
+					'Set the `__nextHasNoMarginBottom` prop to true to start opting into the new styles, which will become the default in a future version',
+			}
+		);
 	}
+	const deprecatedStyles = __nextHasNoMarginBottom
+		? undefined
+		: { marginBottom: 24 };
 
 	return (
 		<div
-			className={ classNames( 'block-editor-line-height-control', {
-				'has-legacy-styles': __unstableHasLegacyStyles,
-			} ) }
+			className="block-editor-line-height-control"
+			style={ deprecatedStyles }
 		>
 			<NumberControl
-				__unstableInputWidth={
-					__unstableHasLegacyStyles ? '60px' : undefined
-				}
+				__unstableInputWidth={ __unstableInputWidth }
 				onChange={ onChange }
 				label={ __( 'Line height' ) }
 				placeholder={ BASE_DEFAULT_VALUE }